// src/constants/enum.ts
export enum CORRECTION_STATUS {
  PENDING = "pending",
  APPROVED = "approved",
  REJECTED = "rejected",
  CANCEL = "cancel",
}

export enum ROLE {
  ADMIN = "admin",
  USER = "user",
}

export enum PUNCH_TYPE {
  SYSTEM = "system",
  CORRECTION = "correction",
}

export enum CORRECTION_ACTION {
  ADD = "add",
  DELETE = "delete",
}

export enum PUNCH_STATUS {
  IN = "in",
  OUT = "out",
}